docs: document change in goldens directory after #35768
The public api .d.ts files were moved from tools/public_api_guard to goldens/public-api. and  and a script
was added in #35768 which assist in testing the current state of the repo against the goldens as
well as a command for accepting all changes to the goldens in a single
command. 

All these changes were not documented added documentation for new script command and an example of the new logs generated.